+++ src/components/Feed/CreateFeed/index.tsx
@@ -1,28 +1,15 @@
 import Image from 'next/image';
-import { File } from 'buffer';
 import classNames from 'classnames/bind';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { useForm, Controller } from 'react-hook-form';
 import axios from 'axios';
 import DefaultButton from '@/components/Common/Buttons/DefaultButton';
-import { CloseIcon, ProfileIcon } from '@/components/Common/IconCollection';
-import { useQuery, useMutation } from '@tanstack/react-query';
-import fetchData from '@/api/fetchData';
-import styels from './CreateFeed.module.scss';
-import { FeedType } from './type';
+import { CloseIcon } from '@/components/Common/IconCollection';
+import { useMutation } from '@tanstack/react-query';
+import { useCreateFeedRequest } from '@/hooks/useCreateFeedRequest';
+import styles from './CreateFeed.module.scss';
+import { FeedType, CreatedFeedTypes, Inputs } from './type';
 import { postFeed } from './api';
-
-interface UrlType {
-  uploadURL: string;
-}
-interface CreatedFeedTypes {
-  profileImage: string | null;
-}
-
-interface Inputs {
-  content: string;
-  feedImage: string[];
-}
 
 /**
  * CreatedFeed component
@@ -31,7 +18,7 @@ interface Inputs {
  */
 
 export default function CreateFeed({ profileImage }: CreatedFeedTypes) {
-  const cn = classNames.bind(styels);
+  const cn = classNames.bind(styles);
   const {
     register,
     handleSubmit,
@@ -43,15 +30,7 @@ export default function CreateFeed({ profileImage }: CreatedFeedTypes) {
   } = useForm<Inputs>();
   const [images, setImages] = useState<Blob[]>([]);
   const [urlBucket, setUrlBucket] = useState<string[]>([]);
-
-  const { refetch: getUrl } = useQuery<UrlType>({
-    queryKey: ['signedUrl'],
-    queryFn: () =>
-      fetchData({
-        param: 'feed/image/create',
-      }),
-    enabled: false,
-  });
+  const { getUrl, deleteImage } = useCreateFeedRequest();
 
   const putUrlMutate = useMutation({
     mutationFn: ({ url, file }: { url: string; file: Blob }) =>
@@ -86,28 +65,6 @@ export default function CreateFeed({ profileImage }: CreatedFeedTypes) {
     putUrlMutate.mutate({ url, file });
   };
 
-  const deleteUrlMutate = useMutation({
-    mutationFn: (url: string) =>
-      axios({
-        method: 'delete',
-        url: `${url}`,
-        headers: {
-          'Access-Control-Allow-Origin': 'https://alpha.cosmo-sns.com',
-        },
-      }),
-    onError: () => console.log('이미지 삭제 요청 에러 '),
-    onSuccess: () => console.log('이미지 삭체 요청 성공'),
-  });
-
-  const deleteImage = (url: string) => {
-    deleteUrlMutate.mutate(url);
-  };
-
-  /**
-   * 제어컴포넌트인 이미지 업로드 input이 onChange 이벤가 일어나면 setImages 세터함수가 실행되어
-   * 컴포넌트가 재랜더링 됩니다. 아래 반환문의 조건부 렌더링이 실행되면서 업로드한 이미지들의 url을 생성해 이미지 프리뷰를
-   * 보여주게 됩니다
-   */
   const updateImageUrls = () => {
     if (images && images.length > 0) {
       let urlList = [];
@@ -150,28 +107,15 @@ export default function CreateFeed({ profileImage }: CreatedFeedTypes) {
 
   const imagePreview = updateImageUrls();
 
-  // const putFileIntoURL = () => {
-  //   urlBucket.map((url, i) => putUrl(url, images[i]));
-  // };
-
-  /**
-   * CloseIcon을 클릭하면 filterImage 함수가 실행됩니다.
-   * @param {number} index - useState images 배열의 index는 이미지를 업로드할때 등록되는 index와 같습니다. useState images 배열을 순회하면서 클릭한 이미지의 index를 제외한 나머지 요소를 반환합니다.
-   */
   const filterImage = (index: number) => {
     const urlBox = getValues('feedImage');
     const filteredImages = images.filter((el, i) => i !== index);
     const filteredUrlBucket = urlBox.filter((el, i) => i !== index);
-
     deleteImage(urlBucket[index]);
     setImages(filteredImages);
     setValue('feedImage', filteredUrlBucket);
     setUrlBucket(filteredUrlBucket);
   };
-
-  // useEffect(() => {
-  //   putFileIntoURL();
-  // }, [urlBucket]);
 
   return (
     <form className={cn('container')} onSubmit={handleSubmit(onSubmit)}>
