+++ src/components/Feed/CreateFeed/type.ts
@@ -2,3 +2,15 @@ export interface FeedType {
   content: string;
   feedImage: string[];
 }
+
+export interface UrlType {
+  uploadURL: string;
+}
+export interface CreatedFeedTypes {
+  profileImage: string | null;
+}
+
+export interface Inputs {
+  content: string;
+  feedImage: string[];
+}
